Fix multiple clone functional test
The test was failing due to a type mismatch in the comparison
(int vs int32).
The destructive lane is failing due to a cdi pod not comping back
up in time. Added some extra logging to determine why

+++ tests/cloner_test.go
@@ -1021,7 +1021,7 @@ var _ = Describe("all clone tests", func() {
 							Expect(err).ToNot(HaveOccurred())
 							restartCount := cloner.Status.ContainerStatuses[0].RestartCount
 							fmt.Fprintf(GinkgoWriter, "INFO: restart count on clone source pod %s: %d\n", clonerPodName, restartCount)
-							Expect(restartCount).To(BeNumerically("<", 2))
+							Expect(restartCount).To(BeNumerically("<", int32(2)))
 						}
 					}
